requirements.txt
openai==0.28
psycopg2-binary
fastapi
uvicorn
pandas
openpyxl
aiofiles